+++ TravellingSalesmanProblemUnitTest/BranchAndBoundTest.cs
@@ -85,7 +85,7 @@ namespace TravellingSalesmanProblemUnitTest
         [Fact]
         public void Scenario4Test()
         {
-            int?[,] grid = { { 0, 10, 15, 20 },
+            double?[,] grid = { { 0, 10, 15, 20 },
                        { 10, 0, 35, 25 },
                        { 15, 35, 0, 30 },
                        { 20, 25, 30, 0 } };
